FEM: fem interface package, rename it to inout
