+++ src/Mod/Fem/femsolver/calculix/tasks.py
@@ -33,8 +33,8 @@ import os.path
 
 import FreeCAD as App
 import FemUtils
-import feminterface.importCcxFrdResults as importCcxFrdResults
-import feminterface.importCcxDatResults as importCcxDatResults
+import feminout.importCcxFrdResults as importCcxFrdResults
+import feminout.importCcxDatResults as importCcxDatResults
 
 from .. import run
 from .. import settings
